ml_project/requirements.txt
antlr4-python3-runtime==4.8
attrs==21.2.0
Bottleneck==1.3.2
certifi==2020.12.5
chardet==4.0.0
confuse==1.4.0
cycler==0.10.0
decorator==4.4.2
Faker==8.1.2
htmlmin==0.1.12
hydra-core==1.0.6
idna==2.10
ImageHash==4.2.0
iniconfig==1.1.1
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
missingno==0.4.2
multimethod==1.4
networkx==2.5.1
numpy==1.20.2
omegaconf==2.0.6
packaging==20.9
pandas==1.2.4
pandas-profiling @ https://github.com/pandas-profiling/pandas-profiling/archive/master.zip
phik==0.11.2
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
tangled-up-in-unicode==0.0.7
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
visions==0.7.1
